Shop system finished.  Inventory nearly finished.
Created a basic shop system that can be added to any NPC, granting the
ability to customize shops per object.  Will add more to it later to
make it look nicer, but for now it is fully functional.  Inventory now
has item changing and item usage, will add equip bar in near future.